server/requirements.txt
flask
flask_socketio
requests
python-dotenv
eventlet
openai
flask-login
openai-agents